consolidate base_query and query, simpler and more transparent to the user
